remove shapes interface in favor of types.ts
