+++ ui/src/components/annotation/annotation.component.tsx
@@ -1,7 +1,7 @@
 // Copyright (C) 2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-import { Annotation as AnnotationInterface } from '../../features/annotator/shapes/interfaces';
+import { Annotation as AnnotationInterface } from '../../features/annotator/types';
 import { ShapeFactory } from './shape-factory.component';
 
 interface AnnotationProps {
@@ -10,7 +10,7 @@ interface AnnotationProps {
 }
 
 export const Annotation = ({ maskId, annotation }: AnnotationProps) => {
-    const { id, color, shape } = annotation;
+    const { id, labels, shape } = annotation;
 
     return (
         <>
@@ -18,10 +18,10 @@ export const Annotation = ({ maskId, annotation }: AnnotationProps) => {
                 mask={maskId}
                 id={`canvas-annotation-${id}`}
                 strokeLinecap={'round'}
-                {...(color !== undefined
+                {...(labels.length > 0
                     ? {
-                          fill: color,
-                          stroke: color,
+                          fill: labels[0].color,
+                          stroke: labels[0].color,
                           strokeOpacity: 'var(--annotation-border-opacity)',
                       }
                     : {})}
