+++ ui/src/features/annotator/shapes/rectangle.component.tsx
@@ -1,7 +1,7 @@
 // Copyright (C) 2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-import { Rect as RectInterface } from './interfaces';
+import { Rect as RectInterface } from '../../annotator/types';
 
 interface RectangleProps {
     rect: RectInterface;
