+++ ui/src/features/annotator/tools/bounding-box-tool/anchor.component.tsx
@@ -5,7 +5,7 @@ import { CSSProperties, PointerEvent, ReactNode, useState } from 'react';
 
 import { isFunction } from 'lodash-es';
 
-import { Point } from '../shapes/interfaces';
+import { Point } from '../../types';
 import { isLeftButton } from './utils';
 
 interface AnchorProps {
