+++ ui/src/features/annotator/tools/bounding-box-tool/translate-shape.component.tsx
@@ -3,8 +3,8 @@
 
 import { PointerEvent, useState } from 'react';
 
-import { Annotation } from '../annotation/annotation.component';
-import { Annotation as AnnotationInterface, Point } from '../shapes/interfaces';
+import { Annotation } from '../../../../components/annotation/annotation.component';
+import { Annotation as AnnotationInterface, Point } from '../../types';
 import { allowPanning, isLeftButton } from './utils';
 
 const STROKE_WIDTH = 2;
