+++ ui/src/features/annotator/types.ts
@@ -1,21 +1,32 @@
 // Copyright (C) 2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-export type BoundingBox = {
-    type: 'bounding-box';
+export interface RegionOfInterest {
     x: number;
     y: number;
     width: number;
     height: number;
-};
+}
 
-export type Point = { x: number; y: number };
-export type Polygon = {
-    type: 'polygon';
-    points: Array<Point>;
-};
+export interface Point {
+    x: number;
+    y: number;
+}
 
-export type Shape = BoundingBox | Polygon;
+export interface Rect {
+    readonly shapeType: 'rect';
+    readonly x: number;
+    readonly y: number;
+    readonly width: number;
+    readonly height: number;
+}
+
+export interface Polygon {
+    readonly shapeType: 'polygon';
+    readonly points: Point[];
+}
+
+export type Shape = Rect | Polygon;
 
 export type Label = { id: string; name: string; color: string; isPrediction: boolean; score?: number };
