+++ ui/tests/test-utils/mocked-annotation.ts
@@ -8,7 +8,7 @@ export const getMockedAnnotation = (annotation?: Partial<Annotation>): Annotatio
     return {
         id: 'annotation-1',
         shape: {
-            type: 'bounding-box',
+            shapeType: 'rect',
             x: 10,
             y: 20,
             width: 100,
